implemented task converter for parsing the correct type from json

+++ Services.Habitica/Model/Daily.cs
@@ -45,7 +45,7 @@ namespace Trellabit.Services.Habitica.Model
         public List<Checklist> Checklist { get; set; }
 
         [JsonProperty("streak")]
-        public double Streak { get; set; }
+        public long Streak { get; set; }
 
         // TODO: everyX int
         // TODO: frequency
